pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "repqa"
version = "0.1.0"
description = "Smartphone-based rep quality analysis for home PT exercises (shoulder rehab)"
requires-python = ">=3.11"
dependencies = [
    "mediapipe>=0.10",
    "ffmpeg-python>=0.2",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "pandas>=2.1",
    "pyarrow>=14.0",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "matplotlib>=3.8",
    "opencv-python>=4.8",
    "numpy>=1.26",
    "fastapi>=0.104",
    "uvicorn>=0.24",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
